fix: move head outside of pages, cleanup
