+++ .nvmrc
@@ -1,1 +1,1 @@
-14.15.0
+18.10.0
