+++ gatsby-config.js
@@ -1,15 +1,14 @@
-require('dotenv').config();
+require("dotenv").config();
 
 module.exports = {
   plugins: [
     `gatsby-plugin-sharp`,
     `gatsby-transformer-sharp`,
     `gatsby-plugin-image`,
-    `gatsby-plugin-react-helmet`,
     `gatsby-plugin-typescript`,
     `gatsby-plugin-styled-components`,
     {
-      resolve: 'gatsby-source-filesystem',
+      resolve: "gatsby-source-filesystem",
       options: {
         name: `pages`,
         path: `${__dirname}/src/pages`,
@@ -24,7 +23,7 @@ module.exports = {
       },
     },
     {
-      resolve: 'gatsby-plugin-web-font-loader',
+      resolve: "gatsby-plugin-web-font-loader",
       options: {
         typekit: {
           id: process.env.TYPEKIT_ID,
