+++ package.json
@@ -34,10 +34,9 @@
     "@fortawesome/react-fontawesome": "^0.1.14",
     "babel-plugin-styled-components": "^2.0.2",
     "dotenv": "^10.0.0",
-    "framer-motion": "^4.0.0",
+    "framer-motion": "^8.0.2",
     "gatsby": "^5.3.3",
     "gatsby-plugin-image": "^3.3.2",
-    "gatsby-plugin-react-helmet": "^5.4.0",
     "gatsby-plugin-sharp": "^5.3.2",
     "gatsby-plugin-styled-components": "^6.3.0",
     "gatsby-plugin-typescript": "^5.3.1",
@@ -47,16 +46,14 @@
     "gatsby-transformer-sharp": "^5.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-helmet": "^6.1.0",
     "react-router-dom": "^5.2.0",
     "smoothscroll-polyfill": "^0.4.4",
-    "styled-components": "^5.2.1",
+    "styled-components": "^5.3.6",
     "webpack": "^5.65.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.10",
-    "@types/react-helmet": "^6.1.4",
     "@types/react-router-dom": "^5.1.7",
     "@types/styled-components": "^5.1.21",
     "typescript": "^4.9.4"
