+++ src/pages/index.tsx
@@ -1,17 +1,8 @@
-import React from 'react';
-import { Helmet } from 'react-helmet';
-import './styles.css';
-import { AnimateSharedLayout, motion } from 'framer-motion';
-import GlobalStyles from '../GlobalStyles';
-import App from '../components';
-
-const seo = {
-  url: 'https://rickyrhodes.dev',
-  title: 'Ricky Rhodes | Software Engineer',
-  description: 'TypeScript, GraphQL, NodeJS, Gatsby',
-  favicon: '',
-  image: "https://dl.dropboxusercontent.com/s/qp7m0pjsrlf3w88/meta.png"
-};
+import React from "react";
+import { AnimateSharedLayout, motion } from "framer-motion";
+import GlobalStyles from "../GlobalStyles";
+import App from "../components";
+import "./styles.css";
 
 const Home = () => {
   return (
@@ -23,8 +14,7 @@ const Home = () => {
         animate={{ opacity: 1, y: 0 }}
         transition={{
           duration: 0.6,
-        }}
-      >
+        }}>
         <AnimateSharedLayout>
           <App />
         </AnimateSharedLayout>
@@ -34,4 +24,4 @@ const Home = () => {
 };
 
 export default Home;
-export { Head } from './utils/head'
+export { Head } from "../lib/head";
